Changes to the UI and Level UP system

+++ Assets/Scripts/PlayerStats.cs
@@ -32,8 +32,8 @@ public class PlayerStats : MonoBehaviour
     public int PlayerLevel;
 
     [Header("Player skills")]
-    public int PlayerLuck; // Chance for critical
-    public int PlayerCritical; // Critical damage multiplier
+    public float PlayerLuck; // Chance for critical
+    public float PlayerCritical; // Critical damage multiplier
     //Special skill: poison
     public int PlayerPoisonDmg; // Poison damage
     public int PlayerPoisonTime; // Poison duration
